% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{cbin}
\alias{cbin}
\title{cluster by 'include next' condition}
\usage{
cbin(incl.next)
}
\arguments{
\item{incl.next}{a logical vector}
}
\value{
an integer valued grouping vector
}
\description{
Given a logical variable indicating if the next line should belong to the
same cluster as the current line, this function creates an integer-valued
clustering variable.
}
