% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{compare two vectors}
\usage{
compare(x, y, verbose = FALSE)
}
\arguments{
\item{x}{input vector to compare to y}

\item{y}{input vector to compare to x}

\item{verbose}{logical, print summary statements}
}
\value{
a logical vector of length 1
}
\description{
compare two vectors, return TRUE if they are missing in the same places and
have the same values in all others
}
