% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_bias.R
\name{coxreg_bias}
\alias{coxreg_bias}
\title{Bias analysis for cox regression}
\usage{
coxreg_bias(
  data,
  surv = NULL,
  main,
  bnry = NULL,
  real = NULL,
  xtra.adj = NULL,
  bnry.manual = NULL,
  real.manual = NULL
)
}
\arguments{
\item{data}{data frame}

\item{surv}{an 'stab' with 1 row or, if surv pair can be identified through
the default affix, the common name (label) for the pair}

\item{main}{name of main effect (binary)}

\item{bnry}{names of binary variables to include}

\item{real}{names of continuous 'normalish' variables to include}

\item{xtra.adj}{extra string to include in formula}

\item{bnry.manual}{additional binary confounders not in model, named list
where each element is HR, proportion at \code{main == 0}, proportion at
\code{main == 1}}

\item{real.manual}{additional real confounders not in model, named list
where each element is HR, mean at \code{main == 0}, mean at \code{main
== 1}}
}
\value{
a data.frame with entries
\describe{
\item{"term"}{The term (variable)}
\item{"type"}{Type of variable: main (the main effect), bnry (binary), real
(continuous). Terms added via 'xtra.adj' will have NA here.}
\item{"manual"}{Indicator for 'manually added' variables, i.e. hypothesized
variables.}
\item{"stat0"}{Mean value for variable at main = 0}
\item{"stat1"}{Mean value for variable at main = 1}
\item{"adjHR (adjHR.l, adjHR.u)"}{The HR (lower, upper CI) in an adjusted
model (terms specified by bnry, real, and xtra.adj). Note: values
are included here for the 'manual' terms although they do not
actually appear in the model. }
\item{"mainHR (mainHR.l, mainHR.u)"}{The HR for main if a confounder with
the parameters specified (stat0, stat1, adjHR) was present.}
\item{"mainHRinv (mainHRinv.l, mainHRinv.u)"}{Similar to 'mainHR' but the
effect of the confounder is inversed}
}
Also, there is an attribute 'tidy' with easy-to-plot information.
}
\description{
Get estimates of 'true' effect given the addition of a confounder, as
calculated by \code{update_HR}.
}
