% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_change.R
\name{coxreg_change}
\alias{coxreg_change}
\title{treatment HR changes}
\usage{
coxreg_change(
  data,
  surv = NULL,
  main,
  terms,
  uni = TRUE,
  full = TRUE,
  inc = FALSE,
  exc = FALSE,
  decr.inc = NULL,
  decr.exc = NULL
)
}
\arguments{
\item{data}{data frame}

\item{surv}{an 'stab' with 1 row or, if surv pair can be identified through
the default affix, the common name (label) for the pair}

\item{main}{name of main effect (binary)}

\item{terms}{vector of terms to adjust for (can be named)}

\item{uni}{logical; include each term's 'univariate' effect on main?}

\item{full}{logical; include each term's 'exclusion from full model'-effect
on main?}

\item{inc}{logical; include each  term's effect, by sequential inclusion, on
main?}

\item{exc}{logical; include each  term's effect, by sequential exclusion, on
main?}

\item{decr.inc}{logical; decreasing order for sequential inclusion?}

\item{decr.exc}{logical; decreasing order for sequential exclusion?}
}
\value{
a list of data frames
}
\description{
examine  changes in  main effect  (HR, Cox  regression) when  covariates are
added one at a time; in solation (both included from minimal model and
excluded from full model), sequentially increasing (starting from
a small model), or sequentially decreasing (starting from large model)
}
