% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{cumsum2}
\alias{cumsum2}
\title{partial cumulative summation}
\usage{
cumsum2(x, n = Inf)
}
\arguments{
\item{x}{numeric vector}

\item{n}{integer; number of elements to sum}
}
\value{
a vector of the same length as x
}
\description{
A version of cumsum that will sum only the previous n (including the
current) elements of a vector. If argument \code{n} is at least as big as
\code{length(x)} then \code{cumsum2(x,n)} and \code{cumsum(x,n)} are
identical.
}
