% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{cumsum_bounded}
\alias{cumsum_bounded}
\title{bounded cumulative summation}
\usage{
cumsum_bounded(x, low = 0, high = Inf, n = Inf)
}
\arguments{
\item{x}{a numeric vector}

\item{low}{numeric; lower bound can be length 1 or equal length as x}

\item{high}{numeric; upper bound can be length 1 or equal length as x}

\item{n}{integer; used as in \code{cumsum2}}
}
\value{
a numeric vector
}
\description{
a version of cumsum that will stay within specified boundaries
}
\examples{
cumsum_bounded(c(0,-1,-100,1,7,-99,1,1,1,-1), low = -50)
cumsum_bounded(c(0,-1,-100,1,7,-99,1,1,1,-1)) ## low = 0 is default
cumsum_bounded(c(0,-1,-100,1,7,-99,1,1,1,-1), high = 5)
}
