% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{event1trunc}
\alias{event1trunc}
\title{truncate data at first event}
\usage{
event1trunc(data, id = "id", event)
}
\arguments{
\item{data}{a data.frame}

\item{id}{character; name of id variable}

\item{event}{character; name of event variable}
}
\description{
Truncate a time dependent data set at first occurence of a given event (per
individual).
}
\examples{
d <- data.frame(id=1,t1=c(0,5,10,15),t2=c(5,10,15,20),foo=c(0,0,1,0))
event1trunc(d, event = "foo")
}
