% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_df.R
\name{expand_df}
\alias{expand_df}
\title{Expand a data.frame through factors}
\usage{
expand_df(data, factors)
}
\arguments{
\item{data}{a data.frame}

\item{factors}{character vector, name of factors in data}
}
\value{
a data.frame
}
\description{
Fill in missing factor combinations within a data frame. Typical use case is
for filling in a grouped calculation where certain factor combinations are
missing due to there being no data, see the examples given.
}
\examples{
d <- data.table(gr = factor(LETTERS[c(2,1,1,2,2)],
                            levels = LETTERS[1:3]),
                foo = factor("b", levels = letters[1:2]),
                x = c(0,5,6,1,2))
r <- d[, .(n = .N, mean = mean(x)), by = .(gr, foo)]
r ## <--  not all factor combinations are present
expand_df(c("gr", "foo"), data = r)
}
