% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_treatment.R
\name{fixed_treatment}
\alias{fixed_treatment}
\alias{fixed_switch}
\alias{onoff_treatment}
\title{treatment periods defined by fixed durations}
\usage{
fixed_treatment(
  data,
  id = "id",
  t = "t",
  state = "state",
  run = "run",
  null.state = "",
  simplify = TRUE
)

fixed_switch(data, null.state = "", simplify = TRUE)

onoff_treatment(data, id = "id", t = "t", run = "run", simplify = TRUE)
}
\arguments{
\item{data}{a data.frame containing 'id', 't', 'state' and, optionally,
'run'.}

\item{id}{character; name of id variable}

\item{t}{character; name of time variable}

\item{state}{character; name of treatment variable}

\item{run}{either name of variable (character) containing treatment
durations, or the numeric value for all treatment durations}

\item{null.state}{value for 'state' when treatment ends}

\item{simplify}{logical; if TRUE keep only minimal information enough to
keep track of changes in treatment.}
}
\value{
a data.frame
}
\description{
Given a data frame ('data') with individuals (identified by 'id'), treatment
('state') initiated (oor continued) at 't' with a fixed duration ('run'), we
calculate the times at which treatment ends and reverts back to the
'null.state' (or changes). Note 1: a treatement can be cut short by the
initiation of a different treatement. Note 2: duplicated times (per
individual) will be ignored.
}
\section{Functions}{
\itemize{
\item \code{fixed_switch()}: The workhorse for fixed_treatment; a rigid
function that requires a data.table with variables named 'id' (which
should be a constant), 't' (which the data.table should be ordered on),
'state', and 'run'. Probably don't use this directly, there are no
checks or warnings of any kind with this function

\item \code{onoff_treatment()}: A wrapper for fixed_treatment where the only
state is 1 (and not explicity in the data) and the null state is 0.

}}
\examples{
df <- data.frame(id = 1,
                 t = c(0, 10, 100, 110),
                 state = LETTERS[c(1, 1, 1, 2)])
fixed_treatment(data = df, run = 50, simplify = TRUE)
fixed_treatment(data = df, run = 50, simplify = FALSE)
}
