% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepit.R
\name{grepit2wide}
\alias{grepit2wide}
\title{widen grepit result}
\usage{
grepit2wide(x, event.only = FALSE)
}
\arguments{
\item{x}{the return object from \code{grepit}}

\item{event.only}{logical; event only data, else time and event components
for each alias}
}
\value{
Wide form of grepit. If \code{event.only=FALSE} each alias has the
match associated with \code{first.id == 1} turned into variables
'ev.alias' (value of 'event') and 't.alias' (value of time). If
\code{event.only=FALSE}, each alias gets a column each with the value of
'event' associated with \code{first.id == 1}
}
\description{
Apply \code{data.table::dcast} to turn the result from grepit into wide format
(one row per {id, begin, end}) with time and event component for each alias.
}
