% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-gtab.R
\name{gtab-fncs}
\alias{gtab-fncs}
\alias{create_gtab}
\alias{check_gtab}
\alias{gtab_equiv2factor}
\alias{gtab2factor}
\title{group table (gtab) functions}
\usage{
create_gtab(
  term,
  data,
  all = FALSE,
  all.first = FALSE,
  all.nm = "All",
  rev = FALSE
)

check_gtab(gtab, n = NULL)

gtab_equiv2factor(gtab, na.ok = FALSE, verbose = TRUE)

gtab2factor(gtab, na.ok = FALSE, verbose = FALSE)
}
\arguments{
\item{term}{character; name of variable in data}

\item{data}{a data.frame}

\item{all}{logical; use all rows as one of the groups}

\item{all.first}{logical; put 'all' first}

\item{all.nm}{character; name of 'all' column}

\item{rev}{logical; reverse order of the levels of 'term}

\item{gtab}{data.frame; a "grouping table" (gtab)}

\item{n}{optional; number of rows wanted}

\item{na.ok}{logical; when considering a gtab as a factor, are missing}

\item{verbose}{logical; print stuff
values permissible?}
}
\description{
group table (gtab) functions
}
\details{
create_gtab: create grouping table from a term in data

check_gtab: check a grouping table

gtab_equiv2factor: test if grouping table could be defined by a
single factor

gtab2factor: create a factor from a grouping table (if possible)
}
