% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute00date.R
\name{impute00date}
\alias{impute00date}
\alias{impute_a_00_date}
\title{fix censored dates}
\usage{
impute00date(x, low.bound = as.Date(NA_character_))

impute_a_00_date(x, low.bound = as.Date(NA_character_))
}
\arguments{
\item{x}{character dates (possibly censored) in format '\%Y\%m\%d'
e.g. "20070101" (if valid), or "20070100" (if censored)}

\item{low.bound}{Date, the lower bound. A vector of the same length as 'x'
(or of length 1)}
}
\description{
Fix censored dates of the type '20110000' or '20110200' by
selecting the midpoint of the censored time interval. If a lower bound
is given, the midpoint between this point and the end of the censored
interval is chosen.
}
\section{Functions}{
\itemize{
\item \code{impute_a_00_date()}: impute_a_00_date: the workhorse for impute00date
(can only work with inputs of length 1)

}}
