% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itc_grep.R
\name{itc_grep}
\alias{itc_grep}
\title{individual time constrained grep}
\usage{
itc_grep(
  pattern,
  x,
  data,
  id,
  date,
  units,
  units.id = id,
  begin = NULL,
  end = NULL,
  include = c(TRUE, TRUE),
  ...,
  data.keep = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pattern}{a vector of search strings (regular expressions) (the names
attribute will be used as alias if it exists)}

\item{x}{names of variables to search in (given in order of importance)}

\item{data}{a data frame}

\item{id}{name of id variable (in 'data')}

\item{date}{name of associated date variable (in 'data')}

\item{units}{a data frame containing id's as well as (but optionally)
'begin' and 'end' variables}

\item{units.id}{variable name in 'units' to use as id (by default the same
as 'id') N.B. a unit can appear several times, and will be identified
alongside 'begin' and 'end' (a soft warning will be given if these 3
variables are not enough for uniqueness)}

\item{begin}{variable name in 'units' to use as begin, if missing will be
set to earliest date in data}

\item{end}{variable name in 'units' to use as end, if missing will be set to
latest date in data}

\item{include}{length 2 logical vector specifying if lower (first entry) and
upper (second entry) bounds are inclusive (\code{TRUE}) or not
(\code{FALSE})}

\item{...}{arguments passed to \code{data.table::like} for identifying matches}

\item{data.keep}{character vector of variables you want to keep from 'data'}

\item{verbose}{if \code{TRUE} the function will give helpful and/or annoying
messages}
}
\value{
A data frame with
\itemize{
\item id the id variable
\item alias the name of pattern searched for (else p1, p2, etc)
\item date the date of assicated match
\item time days from 'begin' to 'date'
\item event indicator for a match
\item begin the begin date (could be individual)
\item end the end date (could be individual)
\item match.in the variable the match was found in
\item match the match found
\item first.id indicator for first occurence of associated
id/begin/end-combination
\item first.id_date indicator for first occurence of associated
id/begin/end- AND date combination
\item pattern the pattern searched for
\item ... variables selected with 'data.keep'.
}

Note that any individual can have more than one match. See the vignette for
examples.
}
\description{
in a dataset with one or more variables (typically containing
text) associated with a date, find matches on those variables for
specific individuals within specifed time frames
}
