% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itc_grep.R
\name{itc_grep2wide}
\alias{itc_grep2wide}
\title{widen itc_grep result}
\usage{
itc_grep2wide(x, event.only = FALSE)
}
\arguments{
\item{x}{the return object from \code{itc_grep}}

\item{event.only}{logical; event only data, else time and event components
for each alias}
}
\value{
Wide form of itc_grep. If \code{event.only=FALSE} each alias has the
match associated with \code{first.id == 1} turned into variables
'ev.alias' (value of 'event') and 't.alias' (value of time). If
\code{event.only=FALSE}, each alias gets a column each with the value of
'event' associated with \code{first.id == 1}
}
\description{
Apply \code{data.table::dcast} to turn the result from itc_grep into wide
format (one row per: id, begin, end) with time and event component for each
alias.
}
