% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsplom.R
\name{jsplom}
\alias{jsplom}
\title{Scatter Plot Matrix data sets}
\usage{
jsplom(x, new.names = NULL, ggcode = TRUE)
}
\arguments{
\item{x}{data frame with the data}

\item{new.names}{character; labels for variables}

\item{ggcode}{logical; print example code (ggplot) for creating the splom?}
}
\value{
list with 2 data frames
}
\description{
splom data (johan style)
}
\details{
Create underlying data sets to create a Scatter Plot Matrix in ggplot
}
\examples{
N  <- 100
x <- data.frame(foo = rnorm(N, 10, 1))
x$bar <- rnorm(N, x$foo-2, 1)
x$baz <- rnorm(N, 2*(x$bar-x$foo), 1)
x$qux <- rnorm(N, 1, 1)
x$quuz <- -x$qux - rnorm(N, 0, 0.1)
jsplom(x, new.names = c("The Foo", "The Bar", "The Baz",
                        "The Qux", "The Quuz"))
}
