% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{locf}
\alias{locf}
\title{last observation carried forward}
\usage{
locf(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a vector of same class and length as that of input x
}
\description{
replace missing values with the last non-missing value
}
