% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxcond.R
\name{proxcond}
\alias{proxcond}
\title{proximity of condition (subset-ish)}
\usage{
proxcond(data, id, ..., prox = c(Inf, Inf))
}
\arguments{
\item{data}{data.frame}

\item{id}{character; name of id variable}

\item{...}{a condition passed to subset}

\item{prox}{length 2 non-negative integer vector; how many lines before and
after (respectively) the condition do you want returned}
}
\description{
This function will return lines close to those identified by the condition
}
\examples{
d <- data.frame(id = rep(1:3, each = 5),
                foo = LETTERS[1:15],
                bar = rep(letters[5:1], 3))
proxcond(d, id = "id", foo=="H")
proxcond(d, id = "id", foo=="H", prox = c(Inf,0))
proxcond(d, id = "id", foo=="H", prox = c(1,1))
proxcond(d, id = "id", foo \%in\% c("G", "H"), prox = c(1,1))
proxcond(d, id = "id", bar == "b", prox = c(0,1))
}
