% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv-variable-helpers.R
\name{surv-identifiers}
\alias{surv-identifiers}
\alias{get_or_test_fix}
\alias{get_or_test_prefix}
\alias{surv_paste}
\alias{surv_t}
\alias{surv_e}
\alias{surv_nm}
\alias{slist_time}
\alias{slist_event}
\alias{check_slist}
\alias{surv_search_string}
\alias{create_slist}
\alias{extract_slist_from_names}
\title{surv variable identifier helpers}
\usage{
get_or_test_fix(fix = NULL)

get_or_test_prefix(prefix = NULL)

surv_paste(s, fix)

surv_t(s)

surv_e(s)

surv_nm(s)

slist_time(sl)

slist_event(sl)

check_slist(sl, nm)

surv_search_string()

create_slist(s)

extract_slist_from_names(nm)
}
\arguments{
\item{fix}{character; a prefix or suffix}

\item{prefix}{logical; if TRUE indicates prefix is used, if FALSE suffix}

\item{s}{character; common name of surv variables (differing only in prefix/suffix)}
}
\description{
surv variable identifier helpers
}
\details{
get_or_test_fix: get the default prefix/suffix, or test a candidate

get_or_test_prefix: get the default logical option for if a prefix
(rather than suffix) should be used, or test a candidate

surv_paste: paste a surv name and a (pre/suf)fix

surv_t: create time-component name from a surv name

surv_e: create event-component name from a surv name

surv_nm: create component names from a surv name

slist_time: extract time component names from surv-list

slist_event: extract event component names from surv-list

check_slist: check a surv-list; i.e. check if the component names
exist in 'nm' (typically the variable names)

surv_search_string: create the search string to find the time and
event components (if systemtatically named accoriding to defaults).

create_slist: create surv-list from surv names

extract_slist_from_names: extract all candidate surv-components from
'nm' (typically the variable names)
}
