% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdc_statistic.R
\name{tdc_statistic}
\alias{tdc_statistic}
\alias{tdc_count}
\title{statistics for time dependent covariates}
\usage{
tdc_statistic(data, var, FUN, tstart = "tstart", tstop = "tstop")

tdc_count(data, tstart = "tstart", tstop = "tstop")
}
\arguments{
\item{data}{a data.frame}

\item{var}{character; name of variable of interest}

\item{FUN}{a function; can return a vector of values. Note: if FUN returns a
named vector, these will be retained in the output}

\item{tstart}{character; name of 'tstart' variable}

\item{tstop}{character; name of 'tstop' variable}
}
\description{
Given a time dependent data set ('data' with variables 'tstart' and 'tstop'
indicating the start and stop times for time intervals) with a covariate
('var'), this function will calculate 'FUN' for all distinct time points
}
\section{Functions}{
\itemize{
\item \code{tdc_count()}: Count individuals in study over time

}}
\examples{
d <- data.frame(id = 1:2, tstart = 0, tstop = c(10,12))
u <- data.frame(id = 1:2, t = c(3,7))
tm <- survival::tmerge(d, d, id = id, tstart = tstart, tstop = tstop)
data <- survival::tmerge(tm, u, id = id, U = tdc(t))
tdc_statistic(data, var = "U", FUN = mean)
tdc_statistic(data, var = "U", FUN = function(x){
    c("min" = min(x), "mean" = mean(x), "max" = max(x))
})
tdc_count(data)
}
