% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{time_conversion}
\alias{time_conversion}
\alias{time_num2date}
\alias{time_date2num}
\title{time conversion}
\usage{
time_num2date(num, ref = NULL)

time_date2num(date, ref = NULL)
}
\arguments{
\item{num}{numeric; numeric representation of time}

\item{ref}{Date or NULL; the reference point for numeric times. If NULL, the
function will check if option "survivalist.ref_date" has been set.}

\item{date}{Date; calendar representation of time}
}
\description{
Functions for moving between numeric and Date represention of time.
}
\details{
time_num2date: numeric to Date representation of time

time_date2num: Date to numeric representation of time
}
