% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{tstart2zero}
\alias{tstart2zero}
\title{set tstart to 0}
\usage{
tstart2zero(data, id = "id", tstart = "tstart", tstop = "tstop")
}
\arguments{
\item{data}{data.frame}

\item{id}{character, name of id variable}

\item{tstart}{character, name of tstart variable}

\item{tstop}{character, name of tstop variable}
}
\description{
Reset individual tstart to begin at 0.
}
