% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_bias.R
\name{update_HR}
\alias{update_HR}
\title{bias adjusted HR}
\usage{
update_HR(HR, expG, s0, s1, type = c("bnry", "real"))
}
\arguments{
\item{HR}{numeric; the treatment HR (to be updated) on the outcome}

\item{expG}{numeric; the confounder HR on the outcome}

\item{s0}{numeric; proportion/mean among controls}

\item{s1}{numeric; proportion/mean among cases}

\item{type}{character; specification of type of unmeasured confounder,
binary ('bnry') or continuous ('real')}
}
\description{
calculate what treatment HR would be if we could adjust for an unmeasured
binary (continuous) confounder U having proportion (mean) s0 and s1 in
the control- and treatment group, respectively, and whose HR on the
outcome is expG. The formulas are from Lin, D.  Y., Psaty, B.  M., &
Kronmal, R.  A.  (1998).  Assessing the sensitivity of regression
results to unmeasured confounders in observational studies. Biometrics,
948-963.
}
