% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill_burden.R
\name{pill_burden}
\alias{pill_burden}
\alias{pill_burden_calculator}
\title{treatment burden defined by pill consumption}
\usage{
pill_burden(
  data,
  id = "id",
  t = "t",
  pills = "pills",
  usage = "usage",
  capacity = "capacity",
  burden = "treatment",
  window = Inf,
  dep.rate = 1,
  breaks = NULL,
  break.labels = NULL,
  simplify = TRUE
)

pill_burden_calculator(
  data,
  burden = "treatment",
  window = Inf,
  dep.rate = 1,
  breaks = NULL,
  break.labels = NULL,
  simplify = FALSE
)
}
\arguments{
\item{data}{a data.frame}

\item{id}{character; name of id variable}

\item{t}{character; name of time variable}

\item{pills}{character; name of variable keeping track of pills added to
inventory}

\item{usage}{either name of variable (character) keeping track of pill
usage, or the numeric value for all pill usage. If given as a name that
does not exist in data, the value 1 will be assigned.}

\item{capacity}{either name of variable (character) keeping track of upper
bound for the size of the pill inventory, or the numeric value for all
pill inventory capacities. If given as a name that does not exist in
data, the value Inf will be assigned.}

\item{burden}{character; 'treatment' for cumulative burden of treatment
(0/1) status, else 'pill' for cumulative burden of pills consumed}

\item{window}{numeric; how long back to accumulate the burden}

\item{dep.rate}{numeric; how fast the burden depreciates. If the cumulative
burden has reached V, each unit of time with no treatment will subtract
from the burden by this value, so that the burden is V-dep.rate,
V-2*dep.rate, etc (not being able to go below 0)}

\item{breaks}{vector of values to cut the burden into (as one would use in
the base function \code{cut}). If no breaks are given, the output will
generally be very long.}

\item{break.labels}{character; if breaks are used, this is the 'labels'
argument for \code{base::cut}}

\item{simplify}{logical; if breaks are given, then simplify = TRUE keep only
minimal information enough to keep track of changes in categorized
burden.}
}
\value{
If no simplification has been asked for you get a  data.frame
\itemize{
\item{"id"}{ id variable}
\item{"t"}{ time points}
\item{"pills"}{ pills added to inventory}
\item{"capacity"}{ the current capacity for the inventory}
\item{"store.begin"}{ inventory at beginning of day; note that the daily
consumption preceeds the need for inventory, so inventory bound (if given)
only applies to inventory at the end of day}
\item{"store.end"}{ inventory at end of day}
\item{"use"}{ how many piss were acutally consumed}
\item{"trt"}{ if any pills consumed, one is on (1) treatment}
\item{"stat"}{ the statistic used (trt or use on treatment dats, -dep.rate
on non-treatment days)}
\item{"cumstat"}{ the cumulative summation of 'stat' using cumsum_bounded}
}
}
\description{
This function is at its core very similar to \code{pill_duration} - please
see the documentation for that function also if things are not clear,
especially for the argument id, t, pills, usage, and capacity. Here we are
only concerned with pills for a single type of drug (unlike
\code{pill_duration} there is no \code{state} or \code{null.state}
argument). The function will look at consumption of over time and calculate
the accumulated burden thereof, either of pills consumed (burden = 'pill')
or days with treatment (burden = 'treatment'). It is possible to set a
window within which to calculate the burden (e.g. window = 30) as well as a
depreciation rate (dep.rate) that determines how fast the accumulated burden
(which is a number) should vanish (day by day) when no pills are consumed.
}
\section{Functions}{
\itemize{
\item \code{pill_burden_calculator()}: pill_burden_calculator: The workhorse for
pill_burden; a rigid function that requires a data.table with variables
named 'id' (which should be a constant), 't' (which the data.table
should be ordered on), 'pills', 'usage' and 'capacity'. Probably don't
use this directly, there are no checks or warnings of any kind with this
function

}}
\seealso{
\code{\link{pill_duration}}
}
