% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-stab.R
\name{stab-fncs}
\alias{stab-fncs}
\alias{default.fix}
\alias{default.prefix}
\alias{default.surv.group.name}
\alias{get_or_test_fix}
\alias{get_or_test_prefix}
\alias{surv_paste}
\alias{surv_t}
\alias{surv_e}
\alias{surv_nm}
\alias{check_stab}
\alias{which_surv_component}
\alias{is_surv_component}
\alias{verify_stab}
\alias{surv_search_string}
\alias{create_stab}
\alias{extract_stab_from_names}
\alias{stab2vtab}
\alias{combine_vs_tab}
\alias{slist2stab}
\title{survival table (stab) functions}
\usage{
default.fix()

default.prefix()

default.surv.group.name()

get_or_test_fix(fix = NULL)

get_or_test_prefix(prefix = NULL)

surv_paste(s, fix, prefix = NULL)

surv_t(s, fix = NULL, prefix = NULL)

surv_e(s, fix = NULL, prefix = NULL)

surv_nm(s, fix = NULL, prefix = NULL)

check_stab(stab)

which_surv_component(term, stab)

is_surv_component(term, stab)

verify_stab(stab, nm, warn = TRUE)

surv_search_string()

create_stab(s)

extract_stab_from_names(nm)

stab2vtab(stab, group.name = NULL)

combine_vs_tab(vtab, stab, group.name = NULL)

slist2stab(sl)
}
\arguments{
\item{fix}{character; a prefix or suffix}

\item{prefix}{logical; if TRUE indicates that a prefix is used, if FALSE
suffix}

\item{s}{character; common name of surv variables (differing only in
prefix/suffix)}

\item{stab}{data.frame, a "survival table" (stab) i.e. a data frame
specifying the surv-pair variables. Must contain columns 'label', 'time'
and 'event'. (Optionally it could contain 'group', especially if one
might want to merge this information with a "variable table" (vtab).)}

\item{term}{character; variable name(s)}

\item{nm}{character vector; typically the names of variables in a data set}

\item{group.name}{character; name of grouping for surv variables}

\item{vtab}{data.frame; a "variable table" (vtab)}

\item{sl}{list; a "survival list" (slist) - an older type of specification
of surv-pair variables not in use}
}
\description{
survival table (stab) functions
}
\details{
default.fix: get default value of 'fix'

default.prefix: get the default prefix/suffix

default.surv.group.name: get the default group name for survival
variables

get_or_test_fix: test (and get default if missing) value of 'prefix'

get_or_test_prefix: test (and get the default if missing) the
logical option for if a prefix (rather than suffix) should be used

surv_paste: paste a surv name and a (pre/suf)fix

surv_t: create time-component name from a surv name

surv_e: create event-component name from a surv name

surv_nm: create component names from a surv name

check_stab: check that correct variables are included

which_surv_component: is 'term' the time- event component of a
surv-variable pair (if not: NA_character_)

is_surv_component: is 'term' a component of a surv-variable pair

verify_stab: check a surv-table; i.e. check if the component names
exist in 'nm' (typically the variable names). Return an stab reduced to
those components that exist in nm.

surv_search_string: create the search string to find the time and
event components (if systematically named accoriding to defaults).

create_stab: create survival table from surv names. Use name attribute
of s as labels if such exist.

extract_stab_from_names: extract all candidate surv-components from
'nm' (typically the variable names)

stab2vtab: convert survival table to variable table

combine_vs_tab: create a variable table by combining a variable
table and a surv table

slist2stab: older versions specified surv-pair variables via a list
(surv-list), this function converts a surv-list to a surv table.
}
