% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2event.R
\name{state2event}
\alias{state2event}
\title{create indicators for state change}
\usage{
state2event(state, by = "id", data, by.ref = FALSE)
}
\arguments{
\item{state}{character; name of variable keeping track of a changing state}

\item{by}{character (vector); name(s) of variable(s) to do calculation 'by'
(typically the id variable)}

\item{data}{a data frame or similar}

\item{by.ref}{logical; if data is data.table then TRUE here means defining
variables by reference}
}
\value{
the input data frame with indicators added (named as the levels of s) for state
changes
}
\description{
For each state in 's', create indicator variables for the state
changes. This might be useful e.g. when setting up data for a multistate
model.
}
\examples{
d <- data.frame(id=1, tstart = 0:7, tstop = 1:8,
                state = factor(LETTERS[c(1,1,2,1,1,3,2,2)],
                               levels = LETTERS[4:1]))
state2event("state", id = "id", data = d)
}
