% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_galore.R
\name{survfit_galore}
\alias{survfit_galore}
\title{'survfit' for several outcomes and groups}
\usage{
survfit_galore(formula, data, surv = NULL, gtab = NULL, ...)
}
\arguments{
\item{formula}{formula, of which only the RHS is used}

\item{data}{data.frame; the data}

\item{surv}{specification of surv components (NULL, character vector, or stab)}

\item{gtab}{data.frame; a grouping table (gtab)}

\item{...}{other arguments for survfit}
}
\value{
a data frame
}
\description{
Iterated application of 'survfitted' to all outcomes specificed by 'surv'
plus an added possibility of adding (possibly overlapping) groups by
supplying a list of indexes in a grouping table (gtab).
}
