% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitted.R
\name{survfitted}
\alias{survfitted}
\title{wrapper for 'survfit'}
\usage{
survfitted(formula, data, ..., time.unit = 1L, keep.factors = TRUE, tp = NULL)
}
\arguments{
\item{formula}{formula passed (to survival::survfit.formula)}

\item{data}{data passed}

\item{...}{other arguments passed}

\item{time.unit}{numerical; scale for time, e.g. set to 365.25 to convert
from days o years}

\item{keep.factors}{logical; keep factors in formula as factors in the
output}

\item{tp}{numerical vector; the time points (tp) for which to calculate at risk
numbers. These will be attached as attribute 'at_risk' to the output.}
}
\value{
a data frame
}
\description{
This wrapper for survival::survfit.formula splits the resulting 'strata'
component into its constituting variables, allows for rescaling and can
calculate numbers at risk.
}
