% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitted.R
\name{survfitted}
\alias{survfitted}
\title{wrapper for 'survfit'}
\usage{
survfitted(formula, data, ..., keep.factors = TRUE)
}
\arguments{
\item{formula}{formula passed (to survival::survfit.formula)}

\item{data}{data passed}

\item{...}{other arguments passed}

\item{keep.factors}{logical; keep factors in formula as factors in the output}
}
\value{
a data frame
}
\description{
This wrapper for survival::survfit.formula only splits the resulting 'strata'
component into its constituting variables
}
