% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_surv.R
\name{transform_surv}
\alias{transform_surv}
\alias{combine_surv}
\alias{rescale_surv}
\alias{truncate_surv}
\alias{Survclass_surv}
\alias{landmark_surv}
\title{transform surv variables}
\usage{
combine_surv(surv, data, id = NULL, nm = surv_nm("Combined"), strip = TRUE)

rescale_surv(surv = NULL, data, FUN, id = NULL, strip = TRUE)

truncate_surv(surv = NULL, data, trunc, id = NULL, strip = TRUE)

Survclass_surv(surv = NULL, data, id = NULL, strip = TRUE)

landmark_surv(
  surv = NULL,
  data,
  landmark,
  id = NULL,
  strip = TRUE,
  reset = TRUE
)
}
\arguments{
\item{surv}{specification of surv components}

\item{data}{the data}

\item{id}{name of id variable (optional)}

\item{nm}{name of time and status component of variable created
(combine_surv)}

\item{strip}{logical, strip everything from data apart from the combined
surv calculated and id (if specified)?}

\item{FUN}{function for rescaling, e.g. function(x) x/365.25 if your time
components are originally measured in days and you want them to be
measured in years (rescale_surv)}

\item{trunc}{numeric, the time to truncate at (truncate_surv)}

\item{landmark}{numeric, the time at which to landmark (landmark_surv)}

\item{reset}{logical; when landmarking, should landmark time be reset to
zero?}
}
\description{
Different functions for transforming surv variables: combine, truncate,
rescale
}
\details{
combine_surv: The combined surv variable will take the time from
the smallest time-component of the composing variables. It will have
status 1 if any of the event components whos corresponding time equals
the smallest time is 1, else 0.

rescale_surv: apply function to time component of surv variables

truncate_surv: truncate surv variables

Survclass_surv: create variables of class 'Surv'

landmark_surv: landmark surv variables
}
